requirements.txt
autopep8==1.7.0
backoff==1.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
Deprecated==1.2.13
environs==9.5.0
googleapis-common-protos==1.56.4
grpcio==1.49.1
idna==3.4
marshmallow==3.18.0
opentelemetry-api==1.11.1
opentelemetry-distro==0.30b1
opentelemetry-exporter-otlp==1.11.1
opentelemetry-exporter-otlp-proto-grpc==1.11.1
opentelemetry-exporter-otlp-proto-http==1.11.1
opentelemetry-instrumentation==0.30b1
opentelemetry-launcher==1.8.0
opentelemetry-propagator-b3==1.11.1
opentelemetry-proto==1.11.1
opentelemetry-sdk==1.11.1
opentelemetry-semantic-conventions==0.30b1
packaging==21.3
protobuf==3.20.1
pycodestyle==2.9.1
pyparsing==3.0.9
python-dotenv==0.21.0
requests==2.28.1
six==1.16.0
toml==0.10.2
typing-extensions==4.3.0
uptrace==1.11.1
urllib3==1.26.12
wrapt==1.14.1
